[Refactor] 인프라 패키지 구조 정리 - external/infra → infrastructure
IssueNum #88
