+++ src/main/java/com/umc/devine/infrastructure/fastapi/FastApiReportClient.java
@@ -2,7 +2,7 @@ package com.umc.devine.infrastructure.fastapi;
 
 import com.umc.devine.domain.report.event.ReportCreatedEvent;
 import com.umc.devine.domain.report.service.command.ReportCommandService;
-import com.umc.devine.global.external.clerk.ClerkApiClient;
+import com.umc.devine.infrastructure.clerk.ClerkApiClient;
 import com.umc.devine.infrastructure.fastapi.dto.FastApiReqDto;
 import com.umc.devine.infrastructure.fastapi.dto.FastApiResDto;
 import lombok.extern.slf4j.Slf4j;
