+++ src/main/java/com/umc/devine/infrastructure/github/GitHubService.java
@@ -1,10 +1,10 @@
-package com.umc.devine.global.external.github;
+package com.umc.devine.infrastructure.github;
 
 import com.umc.devine.domain.auth.exception.AuthException;
 import com.umc.devine.domain.auth.exception.code.AuthErrorCode;
-import com.umc.devine.global.external.clerk.ClerkApiClient;
-import com.umc.devine.global.external.github.dto.GitHubContributionDTO;
-import com.umc.devine.global.external.github.dto.GitHubRepositoryDTO;
+import com.umc.devine.infrastructure.clerk.ClerkApiClient;
+import com.umc.devine.infrastructure.github.dto.GitHubContributionDTO;
+import com.umc.devine.infrastructure.github.dto.GitHubRepositoryDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
